refactor: 再次调整目录结构与小米示例代码保持一致
