+++ src/miio/typedef/property_format.h
@@ -1,0 +1,32 @@
+/**
+ * @par     Copyright (c):
+ *
+ *    Copyright 2019 MIoT,MI
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#ifndef __PROPERTY_FORMAT_H__
+#define __PROPERTY_FORMAT_H__
+
+#define MAX_NUMBER_LENGTH 64
+#define MAX_BOOLEAN_LENGTH 8
+
+enum property_format_t {
+  PROPERTY_FORMAT_UNDEFINED = 0,
+  PROPERTY_FORMAT_BOOLEAN = 1,
+  PROPERTY_FORMAT_STRING = 2,
+  PROPERTY_FORMAT_NUMBER = 3,
+};
+
+#endif /* __PROPERTY_FORMAT_H__ */
